backend/requirements.txt
django>=4.0,<5.0
djangorestframework>=3.12
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.0
python-dotenv>=0.19.2
drf-yasg>=1.21.8
django-cors-headers>=3.14.0